pyproject.toml
[project]
name = "floshy-bot"
version = "0.1.0"
description = "Bot discord performant"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "discord-py>=2.6.4",
    "loguru>=0.7.3",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "riotwatcher>=3.3.1",
]

[dependency-groups]
dev = [
    "black>=25.12.0",
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.10",
]

[project.scripts]
bot = "src.main:main"

# Configuration UV
[tool.uv]
package = true

# Build system
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

# Configuration Hatchling - AJOUTEZ CETTE SECTION
[tool.hatch.build.targets.wheel]
packages = ["src"]  # ← Indique où se trouve votre code

# Configuration MyPy
[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
pretty = true

[[tool.mypy.overrides]]
module = "riotwatcher.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "discord.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "loguru.*"
ignore_missing_imports = true

# Configuration Black
[tool.black]
line-length = 88
target-version = ['py312']
include = '\.pyi?$'

# Configuration Ruff
[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = []

# Configuration Pytest
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--cov=src",
    "--cov-report=term-missing",
    "--cov-report=html",
    "-v",
    "--asyncio-mode=auto"
]
asyncio_mode = "auto"

# Configuration Coverage
[tool.coverage.run]
source = ["src"]
omit = ["tests/*", "**/__pycache__/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if __name__ == .__main__.:",
    "raise AssertionError",
    "raise NotImplementedError",
]